# lib/api-zod/src/generated/types/package.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MetaPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Package {
  id: number;
  name: string;
  description: string;
  price: number;
  duration: number;
  features: string[];
  popular?: boolean;
}