# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MetaPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
  /** @nullable */
  phone?: string | null;
}