# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MetaPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: string;
  priority: string;
  category: string;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  tags?: string[];
  createdAt: string;
  updatedAt?: string;
}