# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MetaPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TaskInput {
  title: string;
  description?: string;
  priority: string;
  category: string;
  dueDate?: string;
  assignedTo?: string;
  tags?: string[];
}